+++ spec/salesforce_ar_sync/salesforce_ar_sync_spec.rb
@@ -1,51 +1,31 @@
 require 'spec_helper.rb'
 
-#for testing our environment variables
-SalesforceArSync.config = Hash.new
-SalesforceArSync.config["ORGANIZATION_ID"] = "123456789123456789"
-SalesforceArSync.config["SYNC_ENABLED"] = true
-
-class Contact < ActiveRecord::Base
-  salesforce_syncable :sync_attributes => {:FirstName => :first_name, :LastName => :last_name, :Phone => :phone_number, :Email => :email_address}
-
-  def phone_number=(new_phone_number)
-    self.phone = new_phone_number
-  end
-
-  def email_address
-    self.email
-  end
-
-  def email_address_changed?
-    email_changed?
-  end
-
-  # Hack for Parsing into the proper Timezone
-  def salesforce_updated_at=(updated_at)
-    updated_at = Time.parse(updated_at) if updated_at.present? && updated_at.kind_of?(String)
-    write_attribute(:salesforce_updated_at, updated_at)
-  end
-end
+# for testing our environment variables
+SalesforceArSync.config = {}
+SalesforceArSync.config['ORGANIZATION_ID'] = '123456789123456789'
+SalesforceArSync.config['SYNC_ENABLED'] = false
 
 # the following hash should match the data in SF.com you are testing against
 sample_outbound_message_hash = {
-  :Id => '003A0000014dbEeIAI',
-  :FirstName => 'Rose',
-  :LastName => 'Gonzalez',
-  :Phone => '(512) 757-6000',
-  :SystemModstamp => '2012-03-26T19:54:50.000Z',
-  :WebId__c => 68836
+  Id: '003A0000014dbEeIAI',
+  FirstName: 'Rose',
+  LastName: 'Gonzalez',
+  Phone: '(512) 757-6000',
+  SystemModstamp: '2012-03-26T19:54:50.000Z',
+  NumberOfPosts__c: 2,
+  WebId__c: 68_836
 }
 
 # this message is missing the Phone field on purpose
-# Salesforce does not include empty fields  in the body
+# Salesforce does not include empty fields in the body
 # of an outbound message
 sample_partial_outbound_message_hash = {
-  :Id => '003A0000014dbEeIAI',
-  :FirstName => 'Rose',
-  :LastName => 'Gonzalez',
-  :SystemModstamp => '2012-03-26T19:54:50.000Z',
-  :WebId__c => 68836
+  Id: '003A0000014dbEeIAI',
+  FirstName: 'Rose',
+  LastName: 'Gonzalez',
+  SystemModstamp: '2012-03-26T19:54:50.000Z',
+  NumberOfPosts__c: 2,
+  WebId__c: 68_836
 }
 
 describe SalesforceArSync, :vcr do
@@ -54,40 +34,42 @@ describe SalesforceArSync, :vcr do
       include ActiveModel::Validations::Callbacks
       extend SalesforceArSync::Extenders::SalesforceSyncable
 
-      salesforce_syncable :salesforce_sync_enabled => false,
-        :sync_attributes => {:FirstName => :first_name, :LastName => :last_name},
-        :async_attributes => ["Last_Login__c", "Login_Count__c"],
-        :default_attributes_for_create => {:password_change_required => true},
-        :salesforce_id_attribute_name => :Id,
-        :web_id_attribute_name  => :WebId__c,
-        :activerecord_web_id_attribute_name => :web_id,
-        :salesforce_sync_web_id => false,
-        :web_class_name => 'Contact',
-        :salesforce_object_name => :salesforce_object_name_method_name,
-        :except => :except_method_name,
-        :save_method => :save_method_name,
-        :sync_inbound_delete => false,
-        :sync_outbound_delete => :outbound_delete_method_name,
-        :unscoped_updates => false,
-        :additional_lookup_fields => {:login => :User_ID_Email__c}
+      salesforce_syncable salesforce_sync_enabled: false,
+                          sync_attributes: { FirstName: :first_name, LastName: :last_name },
+                          async_attributes: %w[Last_Login__c Login_Count__c],
+                          default_attributes_for_create: { password_change_required: true },
+                          salesforce_id_attribute_name: :Id,
+                          web_id_attribute_name: :WebId__c,
+                          activerecord_web_id_attribute_name: :web_id,
+                          salesforce_sync_web_id: false,
+                          web_class_name: 'Contact',
+                          salesforce_object_name: :salesforce_object_name_method_name,
+                          except: :except_method_name,
+                          save_method: :save_method_name,
+                          sync_inbound_delete: false,
+                          sync_outbound_delete: :outbound_delete_method_name,
+                          unscoped_updates: false,
+                          additional_lookup_fields: { login: :User_ID_Email__c },
+                          readonly_fields: %i[NumberOfPosts__c]
     end
 
-    it 'should assign values from the options hash to model attributes' do
+    it 'assigns values from the options hash to model attributes' do
       expect(TestSyncable.salesforce_sync_enabled).to eq(false)
-      expect(TestSyncable.salesforce_sync_attribute_mapping).to eq({"FirstName" => :first_name, "LastName" => :last_name})
-      expect(TestSyncable.salesforce_async_attributes).to eq(["Last_Login__c", "Login_Count__c"])
-      expect(TestSyncable.salesforce_default_attributes_for_create).to eq({:password_change_required => true})
+      expect(TestSyncable.salesforce_sync_attribute_mapping).to eq({ 'FirstName' => :first_name, 'LastName' => :last_name })
+      expect(TestSyncable.salesforce_async_attributes).to eq(%w[Last_Login__c Login_Count__c])
+      expect(TestSyncable.salesforce_default_attributes_for_create).to eq({password_change_required: true})
       expect(TestSyncable.salesforce_id_attribute_name).to eq(:Id)
       expect(TestSyncable.salesforce_web_id_attribute_name).to eq(:WebId__c)
       expect(TestSyncable.activerecord_web_id_attribute_name).to eq(:web_id)
-      expect(TestSyncable.salesforce_web_class_name).to eq("Contact")
+      expect(TestSyncable.salesforce_web_class_name).to eq('Contact')
       expect(TestSyncable.salesforce_object_name_method).to eq(:salesforce_object_name_method_name)
       expect(TestSyncable.salesforce_skip_sync_method).to eq(:except_method_name)
       expect(TestSyncable.salesforce_save_method).to eq(:save_method_name)
       expect(TestSyncable.sync_inbound_delete).to eq(false)
       expect(TestSyncable.sync_outbound_delete).to eq(:outbound_delete_method_name)
       expect(TestSyncable.unscoped_updates).to eq(false)
-      expect(TestSyncable.additional_lookup_fields).to eq({:login => :User_ID_Email__c})
+      expect(TestSyncable.additional_lookup_fields).to eq({login: :User_ID_Email__c})
+      expect(TestSyncable.readonly_fields).to eq(%i[NumberOfPosts__c])
     end
   end
 
@@ -101,10 +83,11 @@ describe SalesforceArSync, :vcr do
   end
 
   describe '.salesforce_update' do
-    it 'should raise an exception if the salesforce id is blank' do
-      expect { Contact.salesforce_update(:Id => '', :FirstName => 'Bob') }.to raise_exception(ArgumentError)
-      expect { Contact.salesforce_update(:FirstName => 'Bob')}.to raise_exception(ArgumentError)
+    it 'raises an exception if the salesforce id is blank' do
+      expect { Contact.salesforce_update(Id: '', FirstName: 'Bob') }.to raise_exception(ArgumentError)
+      expect { Contact.salesforce_update(FirstName: 'Bob') }.to raise_exception(ArgumentError)
     end
+
     it 'looks for records matching salesforce id' do
       sf_id = 1
 
@@ -114,6 +97,7 @@ describe SalesforceArSync, :vcr do
       expect(Contact).to receive(:find_by).with(salesforce_id: sf_id)
       Contact.salesforce_update(Id: sf_id)
     end
+
     it 'looks for records matching salesforce or web id' do
       sf_id = 1
       web_id = 20
@@ -125,6 +109,7 @@ describe SalesforceArSync, :vcr do
       expect(Contact).to receive(:find_by).with(id: web_id)
       Contact.salesforce_update(Id: sf_id, WebId__c: web_id)
     end
+
     it 'looks for records matching salesforce id or with a custom field matching web id' do
       sf_id = 1
       web_id = 20
@@ -160,58 +145,57 @@ describe SalesforceArSync, :vcr do
 
       Contact.salesforce_update(Id: sf_id)
     end
-
   end
 
   describe '.salesforce_id_attribute_name' do
-    it "returns the salesforce Id attribute name" do
+    it 'returns the salesforce Id attribute name' do
       expect(Contact.salesforce_id_attribute_name).to eq(:Id)
     end
   end
 
   describe '.salesforce_sync_web_id?' do
-    it "should default to false" do
+    it 'defaults to false' do
       expect(Contact.salesforce_sync_web_id?).to be_falsey
     end
   end
 
   describe '.salesforce_web_id_attribute_name' do
-    it "should default to 'WebId__c'" do
+    it "defaults to 'WebId__c'" do
       expect(Contact.salesforce_web_id_attribute_name).to eq(:WebId__c)
     end
   end
 
   describe '.salesforce_default_attributes_for_create' do
-    it "should default to an empty hash" do
+    it 'defaults to an empty hash' do
       expect(Contact.salesforce_default_attributes_for_create).to eq({})
     end
   end
 
   describe '#salesforce_object_name' do
-    it "returns the current class name" do
-      expect(Contact.new.salesforce_object_name).to eq("Contact")
+    it 'returns the current class name' do
+      expect(Contact.new.salesforce_object_name).to eq('Contact')
     end
 
-    it "calls a method if one is provided" do
+    it 'calls a method if one is provided' do
       class User < ActiveRecord::Base
         include ActiveModel::Validations::Callbacks
         extend SalesforceArSync::Extenders::SalesforceSyncable
 
-        salesforce_syncable :salesforce_object_name => :custom_name
+        salesforce_syncable salesforce_object_name: :custom_name
 
         def custom_name
-          "CustomUser"
+          'CustomUser'
         end
       end
 
       user = User.new
-      expect(user.salesforce_object_name).to eq("CustomUser")
+      expect(user.salesforce_object_name).to eq('CustomUser')
     end
   end
 
   describe '.ar_sync_inbound_delete?' do
     context 'by default' do
-      it 'should return true' do
+      it 'returns true' do
         contact = Contact.new
 
         expect(contact.ar_sync_inbound_delete?).to be_truthy
@@ -219,23 +203,23 @@ describe SalesforceArSync, :vcr do
     end
 
     context 'sync_inbound_delete set to true' do
-      it 'should return true' do
-        contact = Contact.new(:sync_inbound_delete => true)
+      it 'returns true' do
+        contact = Contact.new(sync_inbound_delete: true)
 
         expect(contact.ar_sync_inbound_delete?).to be_truthy
       end
     end
 
     context 'sync_inbound_delete set to :method_name' do
-      it 'should return the result of the method' do
+      it 'returns the result of the method' do
         class DeleteTest < ActiveRecord::Base
           include ActiveModel::Validations::Callbacks
           extend SalesforceArSync::Extenders::SalesforceSyncable
 
-          salesforce_syncable :sync_inbound_delete => :sync_delete
+          salesforce_syncable sync_inbound_delete: :sync_delete
 
           def sync_delete
-            return true
+            true
           end
         end
 
@@ -247,7 +231,7 @@ describe SalesforceArSync, :vcr do
 
   describe '.ar_sync_outbound_delete?' do
     context 'by default' do
-      it 'should return false' do
+      it 'returns false' do
         contact = Contact.new
 
         expect(contact.ar_sync_outbound_delete?).to be_falsey
@@ -255,23 +239,23 @@ describe SalesforceArSync, :vcr do
     end
 
     context 'sync_outbound_delete set to false' do
-      it 'should return false' do
-        contact = Contact.new(:sync_outbound_delete => false)
+      it 'returns false' do
+        contact = Contact.new(sync_outbound_delete: false)
 
         expect(contact.ar_sync_outbound_delete?).to be_falsey
       end
     end
 
     context 'sync_outbound_delete set to :method_name' do
-      it 'should return the result of the method' do
+      it 'returns the result of the method' do
         class DeleteTest < ActiveRecord::Base
           include ActiveModel::Validations::Callbacks
           extend SalesforceArSync::Extenders::SalesforceSyncable
 
-          salesforce_syncable :sync_outbound_delete => :sync_delete
+          salesforce_syncable sync_outbound_delete: :sync_delete
 
           def sync_delete
-            return false
+            false
           end
         end
 
@@ -283,24 +267,22 @@ describe SalesforceArSync, :vcr do
 
   describe 'salesforce_skip_sync?' do
     context 'by default' do
-      it 'should return false' do
+      it 'returns false' do
+        SalesforceArSync.config['SYNC_ENABLED'] = true
         expect(Contact.new.salesforce_skip_sync?).to be_falsey
+        SalesforceArSync.config['SYNC_ENABLED'] = false
       end
     end
 
     context 'when SYNC_ENABLED is false in the global SalesforceArSync.config hash' do
-      it 'should return true' do
-        SalesforceArSync.config["SYNC_ENABLED"] = false;
-
+      it 'returns true' do
         expect(Contact.new.salesforce_skip_sync?).to be_truthy
-
-        SalesforceArSync.config["SYNC_ENABLED"] = true;
       end
     end
 
     context 'when salesforce_skip_sync is true on an object' do
       it 'returns true' do
-        contact = Contact.new(:salesforce_skip_sync => true)
+        contact = Contact.new(salesforce_skip_sync: true)
 
         expect(contact.salesforce_skip_sync?).to be_truthy
       end
@@ -312,7 +294,7 @@ describe SalesforceArSync, :vcr do
           include ActiveModel::Validations::Callbacks
           extend SalesforceArSync::Extenders::SalesforceSyncable
 
-          salesforce_syncable :salesforce_sync_enabled => false
+          salesforce_syncable salesforce_sync_enabled: false
         end
 
         sync_test = SyncTest.new
@@ -324,10 +306,10 @@ describe SalesforceArSync, :vcr do
           include ActiveModel::Validations::Callbacks
           extend SalesforceArSync::Extenders::SalesforceSyncable
 
-          salesforce_syncable :salesforce_sync_enabled => false
+          salesforce_syncable salesforce_sync_enabled: false
         end
 
-        sync_test = SyncTest.new(:salesforce_skip_sync => true)
+        sync_test = SyncTest.new(salesforce_skip_sync: true)
         expect(sync_test.salesforce_skip_sync?).to be_truthy
       end
     end
@@ -338,10 +320,10 @@ describe SalesforceArSync, :vcr do
           include ActiveModel::Validations::Callbacks
           extend SalesforceArSync::Extenders::SalesforceSyncable
 
-          salesforce_syncable :except => :custom_sync
+          salesforce_syncable except: :custom_sync
 
           def custom_sync
-            return true
+            true
           end
         end
 
@@ -354,14 +336,14 @@ describe SalesforceArSync, :vcr do
           include ActiveModel::Validations::Callbacks
           extend SalesforceArSync::Extenders::SalesforceSyncable
 
-          salesforce_syncable :except => :custom_sync, :salesforce_sync_enabled => false
+          salesforce_syncable except: :custom_sync, salesforce_sync_enabled: false
 
           def custom_sync
-            return false
+            false
           end
         end
 
-        sync_test = SyncTest.new()
+        sync_test = SyncTest.new
         expect(sync_test.salesforce_skip_sync?).to be_truthy
       end
 
@@ -370,57 +352,55 @@ describe SalesforceArSync, :vcr do
           include ActiveModel::Validations::Callbacks
           extend SalesforceArSync::Extenders::SalesforceSyncable
 
-          salesforce_syncable :except => :custom_sync
+          salesforce_syncable except: :custom_sync
 
           def custom_sync
-            return false
+            false
           end
         end
 
-        sync_test = SyncTest.new(:salesforce_skip_sync => true)
+        sync_test = SyncTest.new(salesforce_skip_sync: true)
         expect(sync_test.salesforce_skip_sync?).to be_truthy
       end
     end
   end
 
   describe '#salesforce_attributes_to_set' do
-
     before(:each) do
       @hash = Contact.new.salesforce_attributes_to_set(sample_outbound_message_hash)
     end
 
-    it 'hash should include salesforce_updated_at' do
+    it 'hash includes salesforce_updated_at' do
       expect(@hash[:salesforce_updated_at]).to eq(sample_outbound_message_hash[:SystemModstamp])
     end
 
-    it 'hash should include salesforce_id' do
+    it 'hash includes salesforce_id' do
       expect(@hash[:salesforce_id]).to eq(sample_outbound_message_hash[:Id])
     end
 
-    it 'hash should only include keys and values in the mapping that we have setters for' do
+    it 'hash only includes keys and values in the mapping that we have setters for' do
       expect(@hash[:first_name]).to eq(sample_outbound_message_hash[:FirstName])
       expect(@hash[:last_name]).to eq(sample_outbound_message_hash[:LastName])
       expect(@hash[:phone_number]).to eq(sample_outbound_message_hash[:Phone])
     end
 
-    it 'hash should only include 5 keys' do
-      expect(@hash.keys.count).to eq(5)
+    it 'hash only includes 6 keys' do
+      expect(@hash.keys.count).to eq(6)
     end
 
     it 'removes any attribute mapped to "id" as we do not want to set our primary key' do
-      allow(Contact).to receive(:salesforce_sync_attribute_mapping).and_return({"WebId__c" => "Id"})
+      allow(Contact).to receive(:salesforce_sync_attribute_mapping).and_return({'WebId__c' => 'Id'})
       contact = Contact.new
 
       contact.salesforce_attributes_to_set(sample_outbound_message_hash).tap do |hash|
         expect(hash[:first_name]).to be_nil
       end
     end
-
   end
 
   describe '#salesforce_process_update' do
-    it 'should update and save record with values passed in to hash' do
-      contact = Contact.new(:first_name => "Bob", :last_name => "Smith")
+    it 'updates and save record with values passed in to hash' do
+      contact = Contact.new(first_name: 'Bob', last_name: 'Smith')
       contact.salesforce_skip_sync = true
       contact.save!
       contact.salesforce_process_update(sample_outbound_message_hash)
@@ -432,8 +412,8 @@ describe SalesforceArSync, :vcr do
       expect(contact.salesforce_updated_at).to eq(Time.parse(sample_outbound_message_hash[:SystemModstamp]))
     end
 
-    it 'should nil out any values not specified in the message from Salesforce' do
-      contact = Contact.new(:first_name => "Bob", :last_name => "Smith", :phone => '5195556677')
+    it 'nils out any values not specified in the message from Salesforce' do
+      contact = Contact.new(first_name: 'Bob', last_name: 'Smith', phone: '5195556677')
       contact.salesforce_skip_sync = true
       contact.save!
       contact.salesforce_process_update(sample_partial_outbound_message_hash)
@@ -445,7 +425,7 @@ describe SalesforceArSync, :vcr do
       expect(contact.salesforce_updated_at).to eq(Time.parse(sample_outbound_message_hash[:SystemModstamp]))
     end
 
-    it 'should save with #save! if no save_method is given' do
+    it 'saves with #save! if no save_method is given' do
       class ProcessUpdateTest < ActiveRecord::Base
         include ActiveModel::Validations::Callbacks
         extend SalesforceArSync::Extenders::SalesforceSyncable
@@ -457,12 +437,12 @@ describe SalesforceArSync, :vcr do
       process_update_test.salesforce_process_update
     end
 
-    it 'should use save with the value of save_method if it is provided' do
+    it 'uses save with the value of save_method if it is provided' do
       class ProcessUpdateTest < ActiveRecord::Base
         include ActiveModel::Validations::Callbacks
         extend SalesforceArSync::Extenders::SalesforceSyncable
 
-        salesforce_syncable :save_method => :save_method_name
+        salesforce_syncable save_method: :save_method_name
       end
 
       process_update_test = ProcessUpdateTest.new
@@ -472,27 +452,50 @@ describe SalesforceArSync, :vcr do
   end
 
   describe '#salesforce_attributes_to_update' do
+    let(:contact) do
+      Contact.create(
+        first_name: 'Bob',
+        last_name: 'Smith',
+        phone: '519 555-1212',
+        email: 'bsmith@example.com',
+        number_of_posts: 2,
+        salesforce_skip_sync: true
+      )
+    end
+
     context 'when passing include_all as true' do
+      it 'ignores all fields that are classified as read only' do
+        expect(contact.salesforce_attributes_to_update(true)).not_to have_key('NumberOfPosts__c')
+      end
+
       it 'returns a hash of all attributes and values included in mapping that have getter methods' do
-        contact = Contact.create(:first_name => "Bob", :last_name => "Smith", :phone => "519 555-1212", :email => "bsmith@example.com", :salesforce_skip_sync => true)
-        contact.first_name = "Bill"
+        contact.first_name = 'Bill'
         contact.salesforce_attributes_to_update(true).tap do |hash|
-          expect(hash["FirstName"]).to eq("Bill")
-          expect(hash["LastName"]).to eq("Smith")
-          expect(hash["Email"]).to eq("bsmith@example.com")
+          expect(hash['FirstName']).to eq('Bill')
+          expect(hash['LastName']).to eq('Smith')
+          expect(hash['Email']).to eq('bsmith@example.com')
         end
       end
     end
+
     context 'when passing include_all as false' do
+      it 'ignores all fields that are classified as read only' do
+        contact.first_name = 'Bill'
+        contact.number_of_posts = 3
+
+        expect(contact.salesforce_attributes_to_update(false)).not_to have_key('NumberOfPosts__c')
+      end
+
       it 'returns a hash of changed attributes and values included in mapping that have getter methods' do
-        contact = Contact.create(:first_name => "Bob", :last_name => "Smith", :phone => "519 555-1212", :email => "bsmith@example.com", :salesforce_skip_sync => true)
-        contact.first_name = "Bill"
-        contact.salesforce_attributes_to_update(true).tap do |hash|
-          expect(hash["FirstName"]).to eq("Bill")
+        contact.first_name = 'Bill'
+
+        contact.salesforce_attributes_to_update(false).tap do |hash|
+          expect(hash['FirstName']).to eq('Bill')
         end
       end
     end
   end
+
   describe '#get_activerecord_web_id' do
     context 'no custom web id' do
       it 'returns the id' do
@@ -500,13 +503,14 @@ describe SalesforceArSync, :vcr do
         expect(contact.get_activerecord_web_id).to eq 1
       end
     end
+
     context 'custom web id' do
       it 'returns the id' do
-        contact = Contact.new(id: 1, last_name: "Johnson")
+        contact = Contact.new(id: 1, last_name: 'Johnson')
         allow(Contact).to receive(:salesforce_sync_web_id?).and_return(true)
         allow(Contact).to receive(:activerecord_web_id_attribute_name).and_return(:last_name)
 
-        expect(contact.get_activerecord_web_id).to eq "Johnson"
+        expect(contact.get_activerecord_web_id).to eq 'Johnson'
       end
     end
   end