+++ CHANGELOG.md
@@ -17,3 +17,7 @@
 * Add the ability to map delete messages to any model without modifying the outbound message in Salesforce
 # Version 2.0.2
 * Updated IP address ranges for Salesforce.
+# Version 3.2.0
+* Adds support for Rails 5
+* Added ability to specify readonly fields for salesforce syncable models so those fields don't try to sync back to Salesforce
+* WebID syncing fixed